requirements.txt
fastapi==0.111.1
uvicorn==0.24.0
pydantic==1.10.9
python-dotenv==1.0.0
